+++ playwright/components/dialog/index.ts
@@ -1,15 +1,20 @@
 import type { Page } from "@playwright/test";
 import {
   ALERT_DIALOG,
+  DIALOG,
   DIALOG_TITLE,
   DIALOG_SUBTITLE,
   OPEN_PREVIEW,
   DIALOG_ARIALABEL,
+  DIALOG_WITH_ROLE,
 } from "./locators";
 
 // component preview locators
 const alertDialogPreview = (page: Page) => {
   return page.locator(ALERT_DIALOG);
+};
+const dialog = (page: Page) => {
+  return page.locator(DIALOG);
 };
 const dialogTitle = (page: Page) => {
   return page.locator(DIALOG_TITLE);
@@ -23,11 +28,16 @@ const openPreviewButton = (page: Page) => {
 const dialogAriaLabel = (page: Page) => {
   return page.locator(DIALOG_ARIALABEL);
 };
+const dialogWithRole = (page: Page, role: string) => {
+  return page.locator(DIALOG_WITH_ROLE(role));
+};
 
 export {
   alertDialogPreview,
+  dialog,
   dialogTitle,
   dialogSubtitle,
   openPreviewButton,
   dialogAriaLabel,
+  dialogWithRole,
 };